A few small changes.
I removed the DAC noise "intercept" from the database,
since it was redundant (and possibly confusing) with the pedestals.  Also,
pedestals are now defined with DAC=3000, which is the default value
used by the DAQ.  The only effect of this convention is that the correction
for dac noise is slp*(dac-3000).  This number is in-elegently applied as a
preprocessor definition in TaEvent and in TaADCCalib separately. This number
_must_ be compatible with the parameter used by the DAQ (really the
devices/adc code), so maybe this would be better served to be a database
parameter?

I also added necessary data words used to keep track of the state of
the dithering coils (bmw_*) in genDevTypes, TaDevices, and the control.db
template.

I renamed the "cal" words for both the adc and scaler data to keep
their device name intact (adc1_1_cal, instead of adc_cal1_1).  This was
necessary to use wildcard with ChanList to get data into the pair tree.

Finally, I further modified the timeslot kluge for the new timing board using
Rich's suggestion, to make it compatible with both new and old boards.