Initial check-in of changes to PAN to allow a cheezy munging of real data to simulate effects of dithering, as well as introduce correlations between detectors and monitors.  Future versions of the simulation (contained in TaSimEvent) are to be expected.
- data simulation is turned on through a flag in the database: simtype.
Currently, only simtype = simdit is recognized.

- an abstract event class VaEvent is used throughout PAN.  TaRun creates
a TaEvent for real data, or a TaSimEvent if data simulation is indicated.

- TaSimEvent contains hardwired parameters describing correlations
between monitors and detectors, detector noise, and modulation coils and
monitors.  A modified Decode() routine in TaSimEvent uses those
parameters to munge the fData array.

- The 2-pass analysis ("pan -2") crashes at the very end.  If this is
a new behaviour, it is probably caused by some careless error in my
design of the Event classes, and more specifically in the delete or
Uncreate methods.

- Also, I am apparently accidentally checking in minor changes to
pan_diff.macro and pan_asym.macro.  Hopefully they're an improvement...