I really should have done all this at once.
Added another option when doing a tree variable draw.  Now one can
specify the tree name (as it is defined within the rootfile) with a
"-tree" option.  For example:

   bcm1:ev_num bcm1>5001 -tree R -type prof -title "BCM1 with cut"

This is useful if the variable exists (with the exact same name) in
two different trees, or if you're doing a more complicated draw.. like:

  (1-bcm1/bcm2):(ev_num*3600) abs(bcm1<5e4) -title "What is this?" -tree R

Also cleaned up the method for setting the title on Tree Variable Draws.