Fix a bug in the Standard Analysis... this analysis was including all event data associated with a pariticular bpm into the pair tree, given these lines (e.g. for bpm12):
  f = ChanListFromName ("bpm", "12mx",  fgNO_STATS + fgCOPY);
  fTreeList.insert (fTreeList.end(), f.begin(), f.end());

This didn't just make a evt_bpm12mx[2] branch... but every variable
that had bpm12 in it (including duplicate copies of evt_bpm12x[2] and
evt_bpm12y[2] that are defined a few lines up in InitChanLists()).

Changed the top line to read:
  f = ChanListFromName ("bpm12mx", "",  fgNO_STATS + fgCOPY);

This produces (what I believe to be) the desired result, without the rest.