A few minor changes:
o Comment out the standard out error message when an object is not found in
the rootfile or draws an empty histogram (in favor of drawing the
error message in the current pad).

o If -type scat is indicated for a Tree Draw, set the drawing option to ""
instead of using the "scat" option.  The histogram binning with "scat" is
horrible.

o Change "sample" plots into "summary" plots, when output is to Postscript
file.
