Ok, this is a bit of kludge, but we may have a need apply different extents to different pos burp cuts.  Since posburp is a "multi-cut", the extent is required, by design, to be the same for each item in the list of participating monitors.
Right now, the need is to separate the 4 target-area posburps from the
energy posburp, so I have treated the general problem with an over-specific
solution: I've added a second posburp cut, called Pos_burp_E. Just
like posburp, it takes additional information: a list of monitors
posmonE and thresholds posburpE.  The units for the burp cut is millimeters
(just like for posburp) and there is no requirement that the
monitors listed in the posmonE field actually be related to energy. Its just
a name.

I've debugged this by demonstrating that both posburp and posburpE
can be used independently or together, and get the same results now as
before I changed the code. So, good luck... hope I didn't screw anything up.