Modifications of obtainBMslopes.C
- Introduced kludge to handle Helium slugs 39-41.  This hasn't been tested,
but it was so simple, how could anything go wrong?  Anyway, if it works
as designed, for those particular slugs (this is all hardwired in) the
det*_bpm12x slopes recorded in the slopes tree will actually be the
det*_bpm10x slopes which should be used for these runs.  Thus, the
fix for using bpm10x for energy regression in these runs will be
transparent and robust through the rest of the analysis tree, albeit
possible labeled in a slightly confusing manner.

- Added new leafs to the slopes tree to take advantage of the new words
lastSCev and lostCount from dithering.

-  Event cuts are now compared to the end of the supercycle pulses,
rather than to the start of the next supercycle, which may provide
more useful supercycles.

- One can cut on the lostCount to limit supercycles with >N events cut from
a supercycle.