Modified multirun files to handle a flexible segmenting of dither slopes. Now dither slope files indexed by "Superslug" numbers should be created, using avgSlopes and the superslug defintions obtained from H(e)DitOrganizer. These averaged files will be read in for every run by the TaMakePairFile routine, and it should gracefully handle common errors like data not found or no having a run with no dithering file specified.
TaPairSelector:  modified to signal gracefully if a requested data word
isn't found in the tree.  Installs default value of -1e6 in this case.

HDitOrganizer, HeDitOrganizer:  several utilities for getting the dither
slope cut or the averaged slope filename given a run number or
given an "superslug" index.  I have these in multirun, but maybe they are
better placed in macro?  The average slope filename structure is hardwired
in, so this needs to be modified for your set of average slope files.

avgSlopes.C:  modified to more easily make use of the H(e)DitOrganizer.

TaMakePairFile: makes use of this new scheme.

SlugH.C, SlugHe.C: useful templates for creating average slug files. Some
configuration files are hardwired in, so check these first!