Implements readout of the qweak adc.  I can already see that I could have done this a little differently, so maybe someday this implementation will be changed.
This code assumes 4-subblocks are read out (consistent with assumptions
in existing DAQ libraries). There is, at present, no way to create a
device derived from one of these sub-blocks, which is too bad because
it would be nice to be able to use this "intrinsic" oversampling.

Also, the most general data base rules for creating a derived device
do not work with the qweak integrator, in this version.  Until I figure
out the problem, one must explicity specify offsets to the appropriate
vqwk word (as is always done for the vqwk raw word).