trivially modified Makefile to support OS X builds (by adding a section for the "Darwin" operating system.
The one thing I'm not sure about is that I also added main.o to the compiled objects for libpan.so.  The build crashed on my machine looking for the main routine, but I wasn't sure that should be required for the .so build... if this is wrong, my apologies, and please do revert back to the commented line.